Get legend to work and redce bar chart to one file... breaking the budget stuff
